+++ tsai/callback/core.py
@@ -58,8 +58,17 @@ class TransformScheduler(Callback):
         return f'{self.__class__.__name__}({self.schedule_func})'
 
 # Cell
-class ShowGraph(ShowGraphCallback):
+class ShowGraph(Callback):
     "(Modified) Update a graph of training and validation loss"
+    order,run_valid=65,False
+    names = ['train', 'valid']
+
+    def before_fit(self):
+        self.run = not hasattr(self.learn, 'lr_finder') and not hasattr(self, "gather_preds")
+        if not(self.run): return
+        self.nb_batches = []
+
+    def after_train(self): self.nb_batches.append(self.train_iter)
 
     def after_epoch(self):
         "Plot validation loss in the pbar graph"
@@ -72,7 +81,23 @@ class ShowGraph(ShowGraphCallback):
         y_max = max((max(rec.losses), max(val_losses)))
         margin = (y_max - y_min) * .05
         y_bounds = (y_min - margin, y_max + margin)
-        self.progress.mbar.update_graph([(iters, rec.losses), (self.nb_batches, val_losses)], x_bounds, y_bounds)
+        self.update_graph([(iters, rec.losses), (self.nb_batches, val_losses)], x_bounds, y_bounds)
+
+    def after_fit(self):
+        plt.close(self.graph_ax.figure)
+
+    def update_graph(self, graphs, x_bounds=None, y_bounds=None, figsize=(6,4)):
+        if not hasattr(self, 'graph_fig'):
+            self.graph_fig, self.graph_ax = plt.subplots(1, figsize=figsize)
+            self.graph_out = display(self.graph_ax.figure, display_id=True)
+        self.graph_ax.clear()
+        if len(self.names) < len(graphs): self.names += [''] * (len(graphs) - len(self.names))
+        for g,n in zip(graphs,self.names): self.graph_ax.plot(*g, label=n)
+        self.graph_ax.legend(loc='upper right')
+        if x_bounds is not None: self.graph_ax.set_xlim(*x_bounds)
+        if y_bounds is not None: self.graph_ax.set_ylim(*y_bounds)
+        self.graph_ax.set_title(f'Losses\nepoch: {self.epoch +1}/{self.n_epoch}')
+        self.graph_out.update(self.graph_ax.figure)
 
 ShowGraphCallback2 = ShowGraph